remove DB logic from teos-common
remove rusqlite dependency from teos-common
