+++ teos-common/Cargo.toml
@@ -10,7 +10,6 @@ edition = "2018"
 # General
 hex = { version = "0.4.3", features = [ "serde" ] }
 prost = "0.9"
-rusqlite = { version = "0.29.0", features = [ "bundled", "limits" ] }
 serde = "1.0.130"
 serde_json = "1.0"
 tonic = "0.6"
