+++ teos/Cargo.toml
@@ -27,7 +27,6 @@ home = "0.5.3"
 log = "0.4"
 prost = "0.9"
 rcgen = { version = "0.8", features = ["pem", "x509-parser"] }
-rusqlite = { version = "0.29.0", features = [ "bundled", "limits" ] }
 sqlx = { version = "0.7", features = ["runtime-tokio", "tls-native-tls", "migrate", "any"] }
 serde = "1.0.130"
 serde_json = "1.0"
